Documentacion

+++ clinica-odontologica/clinica-odontologica/src/main/java/com/backend/clinicaodontologica/dto/salida/paciente/PacienteSalidaDto.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 public class PacienteSalidaDto {
 
-    private int id;
+    private Long id;
     private String nombre;
 
     private String apellido;
@@ -16,7 +16,7 @@ public class PacienteSalidaDto {
     public PacienteSalidaDto() {
     }
 
-    public PacienteSalidaDto(int id, String nombre, String apellido, String dni, LocalDate fechaIngreso, DomicilioSalidaDto domicilio) {
+    public PacienteSalidaDto(Long id, String nombre, String apellido, String dni, LocalDate fechaIngreso, DomicilioSalidaDto domicilio) {
         this.id = id;
         this.nombre = nombre;
         this.apellido = apellido;
@@ -25,11 +25,11 @@ public class PacienteSalidaDto {
         this.domicilio = domicilio;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
